+++ theme/template/functions/consts.php
@@ -3,8 +3,6 @@
  * Category names.
  */
 define('NEWS_CATEGORY', 'news');
-define('SPEAKER_SPONSORED', 'sponsored-speakers');
-define('SPEAKER_COMMUNITY', 'community-speakers');
 define('SPONSOR_PLATINUM', 'platinum-sponsors');
 define('SPONSOR_GOLD', 'gold-sponsors');
 define('SPONSOR_SILVER', 'silver-sponsors');
@@ -13,7 +11,6 @@ define('SPONSOR_OPERATION', 'operation-sponsors');
 /**
  * Page names.
  */
-define('SPEAKERS_PAGE', 'speakers');
 define('SPONSORS_PAGE', 'sponsors');
 define('TIMETABLE_PAGE', 'timetable');